Fix race-condition crash issue where HistoryLoader dialog might be null
There is a race-condition when posting to a handler belonging to a different
thread in which the logic in the handler might not have finished executing
by the time we attempt to access data initialized in it.

To correct this issue, we use a FutureTask to ensure the logic in the Runnable
has finished executing before continuing on.