Add FileSelector library project
FileSelector is a library designed to facilitate the ability to browse
the Android filesystem and select a filepath to open or save files.

The code is taken from:

  http://developer.samsung.com/android/technical-docs/Implementing-a-file-selector-dialog

and modified into a library project.  The modifications include, but are not
limited to, adding `fs_` prefixes to all resource ids and various bugfixes.

This code was chosen because it did not have a restrictive license at the time
of copying (2013-09-26), and it has numerous nifty features such as the ability
to create new folders when saving and the ability to set a file filter.