Refactor ShellCommand, add InteractiveShell, remove networklog.sh scripting
Refactored ShellCommand to facilitate use of an interactive shell.  Also reads
stdout as fast as possible in the background, storing output in a buffer for
later retrieval with ShellCommand.readLine() in order to prevent internal
buffers from becoming full.

Added InteractiveShell class for creating and using an interactive shell. As
such, we now no longer need to create a networklog.sh script to batch commands
together. This reduces the number of SuperUser invocations and removes any
potential SELinux problems related to executing a binary in the app's files
directory as su.

In addition, we now prefer system built-in commands over the app's commands
on Android versions equal to or greater than ICE CREAM SANDWICH in order to
avoid potential SELinux issues related to executing binaries in the app's
files directory as su.