Load history from latest timestamp if no entries in range
Instead of not loading any entries at all when there are no entries
within the range specified, start the range from the most recent
entry available.

This prevents the app from appearing as if "data was lost".

TODO: Implement cursor-based logfile loading such that when scrolling
near the limits of the available ranges, incremential logfile loading
occurs so we can continue scrolling through available data in the
logfile instead of just the data that is resident in memory.