Improve CPU resources when ignoring blocked apps
Use improved iptables rules to ignore logging for blocked apps.

Instead of logging all apps and having to parse each and every log message
to determine if a message belongs to a blocked app, we can now simply add
an iptables rule to not provide any log messages for blocked apps!

This means we no longer have to waste precious CPU/battery resources parsing
messages that we don't care about.